# Conflicts:
#   frontend/nginx.conf

+++ prosn/src/main/java/com/ssafy/prosn/dto/SocialLoginResponseDto.java
@@ -12,7 +12,7 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class SocialLoginResponseDto {
-    private Integer oauthId;
+    private Long oauthId;
     private Platform platform;
     private String email;
     private String name;
